Issue #152: Made simple-cipher more approachable by splitting test classes in a more sensible way and making clear which steps they relate to.
