--- exercises/simple-cipher/src/test/java/SimpleCipherTest.java
@@ -1,11 +1,0 @@
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-        RandomKeyCipherTest.class,
-        IncorrectKeyCipherTest.class,
-        SubstitutionCipherTest.class
-})
-public class SimpleCipherTest {
-}
